pyproject.toml
[project]
name = "meta-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.136.0",
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "huggingface-hub>=1.11.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.7.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.14.0",
    "pygithub>=2.9.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "semanticscholar>=0.12.0",
    "sentence-transformers>=5.4.1",
    "shap>=0.51.0",
    "spacy>=3.8.14",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
